:recycle: [Refactor] 폴더명/파일명 수정
컨벤션에 맞춰 수정
